even more refactoring, moving ChatInput and Window from App.js to separate files
